requirements.txt
boto3==1.37.25
botocore==1.37.25
jmespath==1.0.1
python-dateutil==2.9.0.post0
pyyaml==6.0.2
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
